Cambio en la estructura de precios

+++ src/UI/Controller/Admin/FileController.php
@@ -4,32 +4,31 @@ namespace App\UI\Controller\Admin;
 
 use App\Application\Command\DeleteFileCommand;
 use App\Application\Query\FindFileByIdQuery;
-use App\Application\Query\FindAllStationQuery;
+use App\Application\Query\FindAllFilesQuery;
 use App\Framework\Controller\BaseController;
+use App\Shared\Domain\ValueObject\Page;
+use Symfony\Component\HttpFoundation\Request;
 
 class FileController extends BaseController
 {
-    public function list()
+    public function list(Request $request)
     {
-        $list = $this->ask(new FindAllStationQuery());
+        $page = new Page(
+            $request->query->get('page', 1),
+            10
+        );
+        $list = $this->ask(new FindAllFilesQuery($page));
 
         return $this->render('admin/file.list.html.twig', array(
-            'title' => 'Listado de File',
-            'list' => $list
-        ));
-    }
-
-    public function view($id)
-    {
-        $item = $this->ask(new FindFileByIdQuery($id));
-
-        return $this->render('admin/file.item.html.twig', array(
-            'item' => $item
+            'page' => $page,
+            'paginator' => $list
         ));
     }
 
     public function delete($id)
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $this->dispatch(new DeleteFileCommand($id));
 
         return $this->redirectToRoute('admin_file_list');
